norminette, parsing restructured and rgb parsing fixed, depthoffield changed and bug fixes
