+++ Makefile
@@ -6,13 +6,13 @@ NAME = cub3D
 
 SRCS = main.c ./srcs/parsing/file_type.c ./srcs/utils/init.c ./srcs/utils/parsing_utils.c ./srcs/parsing/check_components.c \
 		./srcs/parsing/map_parser.c ./srcs/parsing/map_parser_helper.c ./srcs/game_funcs/game_init.c ./srcs/game_funcs/game_hooks.c \
-		./srcs/game_funcs/get_texture.c ./srcs/game_funcs/fill_map.c ./srcs/utils/casting_utils.c \
-		./srcs/utils/texture_utils.c ./srcs/game_funcs/draw_line.c ./srcs/raycaster/caster.c ./srcs/raycaster/horizontal.c \
-		./srcs/raycaster/vertical.c ./srcs/utils/init_utils.c ./srcs/utils/hook_utils.c ./srcs/utils/render_utils.c \
-		./srcs/raycaster/minimap.c ./srcs/player_funcs/player_movement.c ./srcs/player_funcs/player_rotation.c ./srcs/utils/mini_map_utils.c \
-		./srcs/utils/draw_utils.c ./srcs/utils/math_utils.c ./srcs/sprites/sprites.c ./srcs/sprites/sprite_utils.c \
-		./srcs/sprites/luffy_utils.c ./srcs/sprites/luffy_set_values.c ./srcs/utils/rgb_utils.c ./srcs/sprites/zorro_utils.c \
-		./srcs/sprites/zorro_set_values.c ./srcs/sprites/draw_sprites.c ./srcs/cleanup/del_tex.c ./srcs/cleanup/free.c
+		./srcs/game_funcs/get_texture.c ./srcs/game_funcs/fill_map.c ./srcs/utils/texture_utils.c ./srcs/game_funcs/draw_line.c \
+		./srcs/raycaster/caster.c ./srcs/raycaster/horizontal.c ./srcs/raycaster/vertical.c ./srcs/utils/init_utils.c \
+		./srcs/utils/hook_utils.c ./srcs/utils/render_utils.c ./srcs/raycaster/minimap.c ./srcs/player_funcs/player_movement.c \
+		./srcs/player_funcs/player_rotation.c ./srcs/utils/mini_map_utils.c ./srcs/utils/draw_utils.c ./srcs/utils/math_utils.c \
+		./srcs/sprites/sprites.c ./srcs/sprites/sprite_utils.c ./srcs/sprites/luffy_utils.c ./srcs/sprites/luffy_set_values.c \
+		./srcs/parsing/rgb_parsing.c ./srcs/sprites/zorro_utils.c ./srcs/sprites/zorro_set_values.c ./srcs/sprites/draw_sprites.c \
+		./srcs/cleanup/del_tex.c ./srcs/cleanup/free.c ./srcs/parsing/texture_path_parsing.c
 
 OBJS = $(SRCS:.c=.o)
