+++ maps/test.cub
@@ -3,30 +3,33 @@ NO ./textures/wooden_wall.png
 SO ./textures/wooden_wall.png
 WE ./textures/wooden_wall_no_window.png
 
-C 225,30,0
-F 225,30,0
+C 255 	 	 , 	 	0 	 	 ,	 	 0
+F 0 	 	 , 	 	0 	 	 	 	 	,    		255
 
-       11111111111111111111111111111
-       10000000010000000000000010001
-       1000S00001000000000000001001
-       100000000D00000000000001001
-       1000000001000000000000010001
-       1000000001111111111111110001
-       1000000000000000000Z00111111
-11111111111111111111110000001
-                      10000001
-                       10000001
-1111111111111111111111100000001
-11111100000000000000000000110001
-11111001111111111000001111110001
-1111001111111111110000011110001
-1110011111111111111000000000001
-1100011111111111111111111111111
-1100111
-1100111
-1100111
-1000001
-1004001
-100L001
-1000001
-1111111
+11111111111111111
+10000000000000001
+10000000000000001
+10000000000000001
+10000000000000001
+10000000000000001
+10000000N00000001
+10000000000000001
+10000000000000001
+10000000000000001
+10000000000030001
+10000000000000001
+11111011111111111
+    101
+11111011111111111
+10000000000000001
+10000000002000001
+10000000000000001
+10200000000200001
+10000000000000001
+10000000000200001
+10200000000000001
+10000000400200001
+10000000000000001
+1000000Z000200001
+1020000000L000001
+11111111111111111
