+++ srcs/game_funcs/get_texture.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 14:16:44 by skunert           #+#    #+#             */
-/*   Updated: 2023/11/02 19:19:27 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/03 18:27:49 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
