+++ srcs/parsing/rgb_parsing.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/25 15:34:30 by skunert           #+#    #+#             */
-/*   Updated: 2023/11/02 19:14:41 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/03 22:35:55 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,13 +59,33 @@ int	check_rgb_loop(char **rgb, char *trmd, int i, int j)
 	while (rgb[i][++j])
 		if (ft_isdigit(rgb[i][j]) == 0 && rgb[i][j] != '\n')
 			return (free_arr(rgb), free(trmd),
-			printf("Error\nno valid rgb nbr\n"), 1);
-	if (ft_strlen(rgb[i]) > 3)
-	{
-		free_arr(rgb);
-		return (free(trmd), printf("Error\nwrong rgb size\n"), 1);
-	}
+				printf("Error\nno valid rgb nbr\n"), 1);
 	return (0);
+}
+
+char	*trim_spaces(char *line)
+{
+	char	*trmd;
+	int		i;
+	int		j;
+	int		count;
+
+	i = -1;
+	j = 0;
+	count = 0;
+	while (line[++i])
+		if (is_whitespace(line[i]) == false)
+			count++;
+	trmd = ft_calloc(count + 1, sizeof(char));
+	if (!trmd)
+		return (NULL);
+	i = -1;
+	while (line[++i])
+	{
+		if (is_whitespace(line[i]) == false)
+			trmd[j++] = line[i];
+	}
+	return (trmd);
 }
 
 int	check_rgb(t_map *init, char *line)
@@ -75,15 +95,21 @@ int	check_rgb(t_map *init, char *line)
 	int		i;
 	int		j;
 
-	i = -1;
-	trmd = ft_strtrim(line, "\n");
-	rgb = ft_split(trmd + 2, ',');
+	i = 1;
+	trmd = trim_spaces(line);
+	if (!trmd)
+		return (1);
+	rgb = ft_split(trmd + 1, ',');
+	if (!rgb)
+		return (free(trmd), 1);
 	if (ft_matrixlen(rgb) != 3)
 		return (free_arr(rgb), free(trmd), printf("Error\nrgb format\n"), 1);
+	i = -1;
 	while (rgb[++i])
 	{
 		j = -1;
-		check_rgb_loop(rgb, trmd, i, j);
+		if (check_rgb_loop(rgb, trmd, i, j) == 1)
+			return (1);
 	}
 	if (get_rgb(init, trmd, rgb) == 1)
 		return (free_arr(rgb), free(trmd), 1);
