+++ srcs/raycaster/caster.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 17:08:51 by njantsch          #+#    #+#             */
-/*   Updated: 2023/11/02 21:06:12 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/03 19:01:20 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
