+++ srcs/utils/draw_utils.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 18:45:04 by skunert           #+#    #+#             */
-/*   Updated: 2023/10/23 17:43:16 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/03 21:08:41 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ void	draw_sky(t_game *g)
 	int	x;
 
 	y = -1;
-	while (++y < 40)
+	while (++y < 38)
 	{
 		x = -1;
 		while (++x < 120)
