+++ srcs/utils/math_utils.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 18:46:47 by skunert           #+#    #+#             */
-/*   Updated: 2023/10/24 14:28:25 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/03 19:33:01 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
